create a component
